+++ sources/@roots/bud-terser/src/bud.env.ts
@@ -1,0 +1,12 @@
+import {Extension} from './'
+import {terser} from './terser.api'
+
+declare module '@roots/bud-framework' {
+  interface Framework {
+    terser: terser
+  }
+
+  interface Modules {
+    '@roots/bud-terser': Extension
+  }
+}
