+++ sources/@roots/bud-terser/src/index.ts
@@ -4,17 +4,28 @@
 /**
  * Adds terser minification support to Bud
  *
- * @see https://roots.io/bud
+ * @see https://rootss.io/bud
  * @see https://github.com/roots/bud
- *
- * @remarks
- * - 💁 Composable - Build exceptional web applications using a modular, hackable build system
- *
- * - 💪 Modern - Modern framework that scales from a single file to thousands of lines of code
- *
- * - 🌱 Easy - Low bundle size and fast build times with little to no configuration
- *
+ * 
  * @packageDocumentation
  */
+
+import './bud.env'
+
+import {Extension, Framework} from '@roots/bud-framework'
+import TerserPlugin from 'terser-webpack-plugin/types'
+
+export type Options = TerserPlugin.BasePluginOptions
+
+export type Plugin = TerserPlugin
+
+export interface Extension
+  extends Extension.CompilerPlugin<
+    TerserPlugin,
+    TerserPlugin.BasePluginOptions
+  > {
+  name: '@roots/bud-terser'
+  options: (app: Framework) => Options
+}
 
 export {name, api, options, boot} from './terser.extension'
