+++ sources/@roots/bud-terser/tsconfig.json
@@ -7,7 +7,6 @@
     "outDir": "./lib/cjs",
     "types": ["node", "@roots/bud-api"]
   },
-  "files": ["./src/bud.env.d.ts"],
   "include": ["src/**/*.ts", "src/**/*.json"],
   "references": [
     {
